Created Modal Component
